feat(cli): refactor update-exports command into modular structure
Refactored the `update-exports` command to a modular structure by separating configuration, utilities, and command logic into distinct files. This improves maintainability, readability, and scalability of the codebase.
