+++ scripts/cli/src/commands/update-exports/index.ts
@@ -1,0 +1,1 @@
+export { createUpdateExportsCommand } from "@/commands/update-exports/command";
